test(ui): split task-suggestion pane coverage
